Extract formatter method in own class
